+++ src/main/java/com/global/api/entities/enums/Environment.java
@@ -2,5 +2,6 @@ package com.global.api.entities.enums;
 
 public enum Environment {
     TEST,
-    PRODUCTION
+    PRODUCTION,
+    QA
 }